@@ -8,27 +8,25 @@
 	"license": "MIT",
 	"type": "project",
 	"require": {
-		"php": ">=7.2.5",
-		"fideloper/proxy": "^4.4",
-		"fruitcake/laravel-cors": "^2.0",
-		"doctrine/dbal": "2.5.13",
+		"php": ">=8.1.0",
+		"doctrine/dbal": "^3.0",
 		"guzzlehttp/guzzle": "^7.0.1",
 		"intervention/image": "^2.7",
-		"laravel/framework": "^8.0",
+		"laravel/framework": "^10.0",
 		"laravel/tinker": "^2.0",
 		"maxmind-db/reader": "1.4.1",
 		"namshi/jose": "7.0.*",
-		"symfony/lock": "^4.0",
-		"tymon/jwt-auth": "^1.0",
-		"laravel/ui": "^3.0"
+		"symfony/lock": "^6.0",
+		"tymon/jwt-auth": "^2.0",
+		"laravel/ui": "^4.0"
 	},
 	"require-dev": {
-		"facade/ignition": "^2.3.6",
+		"spatie/laravel-ignition": "^2.0",
 		"fakerphp/faker": "^1.9.1",
 		"imanghafoori/laravel-microscope": "^1.0",
 		"mockery/mockery": "^1.3.1",
-		"nunomaduro/collision": "^5.0",
-		"phpunit/phpunit": "^9.0"
+		"nunomaduro/collision": "^7.0",
+		"phpunit/phpunit": "^10.0"
 	},
 	"autoload": {
 		"classmap": [
